Added unit test

+++ orders/models.py
@@ -5,7 +5,7 @@ from customers.models import Customer
 class Order(models.Model):
 
     item = models.CharField(max_length=255, blank=False)
-    amount = models.DecimalField(max_digits=5, decimal_places=2)
+    amount = models.IntegerField()
     customer = models.ForeignKey(to=Customer, on_delete=models.CASCADE)
     created_at = models.DateTimeField(auto_now_add=True)
